+++ src/cli/index.ts
@@ -2,12 +2,25 @@
 import { Command } from "commander";
 import { resolveLang } from "../i18n.js";
 import { injectDefaultCommandIfMissing } from "./lib/argv.js";
+
 import { registerGenerate } from "./commands/generate.js";
 import { registerShow } from "./commands/show.js";
 import { registerCheck } from "./commands/check.js";
+import { registerInit } from "./commands/init.js";
 
+// default behavior: no subcommand -> generate
 injectDefaultCommandIfMissing(process.argv, {
-  known: new Set(["generate", "show", "check", "-h", "--help", "-V", "--version", "--lang"]),
+  known: new Set([
+    "generate",
+    "show",
+    "check",
+    "init",
+    "-h",
+    "--help",
+    "-V",
+    "--version",
+    "--lang",
+  ]),
   defaultCommand: "generate",
 });
 
@@ -20,8 +33,11 @@ program
   .showSuggestionAfterError()
   .option("--lang <lang>", "CLI language (en | ru)");
 
-registerGenerate(program, () => resolveLang(program.opts().lang));
-registerShow(program, () => resolveLang(program.opts().lang));
-registerCheck(program, () => resolveLang(program.opts().lang));
+const getLang = () => resolveLang(program.opts().lang);
+
+registerGenerate(program, getLang);
+registerShow(program, getLang);
+registerCheck(program, getLang);
+registerInit(program, getLang);
 
 program.parse(process.argv);